Finalize type-driven UI module split and import fixes
