+++ lib/js/components/layouts/type-tools-grid.mjs
@@ -75,11 +75,11 @@ import {
     _BaseTypeDrivenContainerComponent
   , createTypeToUIElementFunction
   , require
+  , UITypeDrivenContainer
 } from '../type-driven-ui-basics.mjs';
 
 import {
     genericTypeToUIElement
-  , UITypeDrivenContainer
 } from '../type-driven-ui.mjs';
 
 import {
