+++ lib/js/components/type-driven-ui-basics.mjs
@@ -24,6 +24,10 @@ import {
 } from './basics.mjs';
 
 import {
+    StaticTag
+} from './generic.mjs';
+
+import {
     getRegisteredPropertySetup
 } from './registered-properties.mjs';
 
@@ -428,6 +432,61 @@ export class UITypeDrivenList extends _UIBaseList {
 
     _createNewItem(targetPath, insertPosition, items/*, value*/) {
         return this._ItemModel.createPrimalDraft(items.dependencies);
+    }
+}
+
+export class UITypeDrivenContainer extends _BaseTypeDrivenContainerComponent {
+    constructor(widgetBus, _zones, injectable, ppsMap, label) {
+        const localZoneElement = widgetBus.domTool.createElement('div', {'class': 'ui_generic_struct_container'})
+          , zones = new Map([..._zones, ['local', localZoneElement]])
+          ;
+        widgetBus.insertElement(localZoneElement);
+        super(widgetBus, zones);
+
+        const entry = this.widgetBus.getEntry(this.widgetBus.rootPath)
+          , TypeClass = entry instanceof _AbstractDynamicStructModel
+                ? entry.WrappedType
+                : entry.constructor
+          , widgets = this._defineWidgets(TypeClass, injectable, ppsMap, label)
+          ;
+        this._initWidgets(widgets);
+    }
+    _defineWidgets(TypeClass, injectable, ppsMap, label) {
+        const labelDefinition = [
+                {zone: 'local'}
+              , []
+              , StaticTag
+              , 'h4'
+              , {}
+              , [label]
+            ];
+        return [
+            ...(label ? [labelDefinition] : [])
+          , ...this._defineGenericWidgets(
+                    TypeClass
+                  , fieldName=>TypeClass.fields.has(fieldName)
+                  , {zone: 'local'}
+                  , ppsMap
+                  , injectable
+            )
+        ];
+    }
+}
+
+// UITypeDrivenListWithAddButton: wraps UITypeDrivenList with an "add" button.
+// Subclass per use-case to set ADD_BUTTON_TEXT.
+export class UITypeDrivenListWithAddButton extends UITypeDrivenList {
+    static ADD_BUTTON_TEXT = '+ add';
+    constructor(widgetBus, _zones, injectable, label=null) {
+        super(widgetBus, _zones, injectable, label);
+        const addButton = widgetBus.domTool.createElement('button'
+                , {'class': 'ui_type_driven_list-add_button'}
+                , this.constructor.ADD_BUTTON_TEXT);
+        addButton.addEventListener('click', ()=>{
+            const rootPath = Path.fromString(this.widgetBus.getExternalName('collection'));
+            this._create(rootPath, 'insert', null);
+        });
+        this.element.append(addButton);
     }
 }
