+++ lib/js/components/type-driven-ui.mjs
@@ -3,16 +3,12 @@
 import {
     _AbstractEnumModel
   , _AbstractNumberModel
-  , _AbstractDynamicStructModel
   , StringModel
   , BooleanModel
-  , Path
 } from '../metamodel.mjs';
 
 import {
-    StaticTag
-  //, StaticNode
-  , UILineOfTextInput
+    UILineOfTextInput
   , UISelectInput
   , UIToggleButton
   , UILineOfTextOrEmptyInput
@@ -24,8 +20,7 @@ import {
 
 import {
     require
-  , _BaseTypeDrivenContainerComponent
-  , UITypeDrivenList
+  , UITypeDrivenContainer
   , createTypeToUIElementFunction
 } from './type-driven-ui-basics.mjs';
 
@@ -42,7 +37,6 @@ import {
   , TemplateRulesModel
 } from './actors/videoproof-contextual-models.mjs';
 
-
 import {
     OpenTypeFeaturesModel
 } from './actors/models.mjs';
@@ -53,6 +47,7 @@ import {
 
 import {
     UICharGroupContainer
+  , UICharGroupsList
 } from './ui-char-groups.mjs';
 
 import {
@@ -90,7 +85,6 @@ import {
   , UILeadingAlgorithm
 } from './type-spec-fundamentals.mjs';
 
-
 import {
     UIMargins
 } from './ui-margins.typeroof.jsx';
@@ -99,6 +93,7 @@ import {
     UIContextualTemplateContainer
   , UIContextualTemplateRulesList
   , UICharsSelectorContainer
+  , UICharsSelectorItemsList
 } from './ui-contextual-template.mjs';
 
 import {
@@ -196,90 +191,6 @@ export {ProcessedPropertiesSystemMap, ProcessedPropertiesSystemRecord};
  *      fullKey = `${propertyRoot}${fieldName}`
  *      getRegisteredPropertySetup(fullKey, {label:fieldName}).label || fieldName;
  */
-export class UITypeDrivenContainer extends _BaseTypeDrivenContainerComponent {
-    constructor(widgetBus, _zones, injectable, ppsMap, label) {
-        // run super first, so we can use `this` in the widgets definition.
-        const localZoneElement = widgetBus.domTool.createElement('div', {'class': 'ui_generic_struct_container'})
-          , zones = new Map([..._zones, ['local', localZoneElement]])
-          ;
-        // When using StaticNode via widgets, it's not inserted right away.
-        // and the position is lost relative to the sibling widgets to the
-        // end of the container.
-        // zones.get('main').append(localZoneElement);
-        widgetBus.insertElement(localZoneElement);
-        super(widgetBus, zones);
-
-        const entry =  this.widgetBus.getEntry(this.widgetBus.rootPath)
-        const TypeClass = entry instanceof _AbstractDynamicStructModel
-            ? entry.WrappedType
-            : entry.constructor
-            ;
-        const widgets = this._defineWidgets(TypeClass, injectable, ppsMap, label);
-        this._initWidgets(widgets);
-    }
-    _defineWidgets(TypeClass, injectable, ppsMap, label) {
-        const labelDefinition = [
-                {zone: 'local'}
-              , []
-              , StaticTag
-              , 'h4'
-              , {}
-              , [label]
-            ];
-        return [
-            // optional label
-            ...(label ? [labelDefinition] : [])
-          , ...this._defineGenericWidgets(
-                    TypeClass
-                  , fieldName=>TypeClass.fields.has(fieldName) // basically all allowed
-                  , {zone: 'local'}
-                  , ppsMap
-                  , injectable
-            )
-        ];
-    }
-}
-
-class UITypeDrivenListWithAddButton extends UITypeDrivenList {
-    static ADD_BUTTON_TEXT = '+ add';
-    constructor(widgetBus, _zones, injectable, label=null) {
-        super(widgetBus, _zones, injectable, label);
-        const addButton = widgetBus.domTool.createElement('button'
-                , {'class': 'ui_type_driven_list-add_button'}
-                , this.constructor.ADD_BUTTON_TEXT);
-        addButton.addEventListener('click', ()=>{
-            const rootPath = Path.fromString(this.widgetBus.getExternalName('collection'));
-            this._create(rootPath, 'insert', null);
-        });
-        this.element.append(addButton);
-    }
-}
-
-class UICharGroupsList extends UITypeDrivenListWithAddButton {
-    static ADD_BUTTON_TEXT = '+ add char group';
-}
-
-class UICharsSelectorItemsList extends UITypeDrivenListWithAddButton {
-    static ADD_BUTTON_TEXT = '+ add selector';
-
-    _createWrapper(rootPath) {
-        const childWidgetBus = this._childrenWidgetBus
-          , settings = {
-                rootPath
-              , zone: this._childrenMainZone
-            }
-          , dependencyMappings = [
-                ['./', 'value']
-            ]
-          ;
-        return this._initWrapper(childWidgetBus, settings, dependencyMappings
-              , UICharsSelectorContainer  // Constructor
-              , this._zones               // _zones
-              , this._injectable          // injectable
-        );
-    }
-}
-
 export const uiElementsMap = new Map([
             [StringModel, [UILineOfTextInput
                     , require('settings:internalPropertyName', 'value')
