+++ lib/js/components/ui-char-groups.mjs
@@ -14,6 +14,7 @@ import {
 
 import {
     _BaseTypeDrivenContainerComponent
+  , UITypeDrivenListWithAddButton
   , require
 } from './type-driven-ui-basics.mjs';
 
@@ -72,7 +73,6 @@ export function getExtendedChars(charGroups, chars) {
     }
     return extendedMap;
 }
-
 
 // This is an important method! it explains as well, how the
 // `null` comes into the `extendedChars` as a key
@@ -526,7 +526,6 @@ export const
     )
   ;
 
-
 function _truncateSummaryText(value, maxLength = 24) {
     const text = `${value ?? ''}`;
     if(text.length <= maxLength)
@@ -702,4 +701,8 @@ export class UICharGroupContainer extends _BaseTypeDrivenContainerComponent {
             this._activationTestCache = null;
         }
     }
+}
+
+export class UICharGroupsList extends UITypeDrivenListWithAddButton {
+    static ADD_BUTTON_TEXT = '+ add char group';
 }
