+++ lib/js/components/ui-contextual-template.mjs
@@ -20,6 +20,7 @@ import {
 import {
     _BaseTypeDrivenContainerComponentMixin
   , UITypeDrivenList
+  , UITypeDrivenListWithAddButton
 } from './type-driven-ui-basics.mjs';
 
 import {
@@ -300,9 +301,30 @@ class UIContextualTemplateRulesList extends UITypeDrivenList {
     static UIItem = UIContextualTemplateRuleListItem;
 }
 
+class UICharsSelectorItemsList extends UITypeDrivenListWithAddButton {
+    static ADD_BUTTON_TEXT = '+ add selector';
+
+    _createWrapper(rootPath) {
+        const childWidgetBus = this._childrenWidgetBus
+          , settings = {
+                rootPath
+              , zone: this._childrenMainZone
+            }
+          , dependencyMappings = [
+                ['./', 'value']
+            ]
+          ;
+        return this._initWrapper(childWidgetBus, settings, dependencyMappings
+              , UICharsSelectorContainer
+              , this._zones
+              , this._injectable
+        );
+    }
+}
 
 export {
     UIContextualTemplateContainer,
     UIContextualTemplateRulesList,
     UICharsSelectorContainer,
+    UICharsSelectorItemsList,
 };
